create new quoter using new dex params

+++ test/mantle/utils/CommonAddresses.f.sol
@@ -105,11 +105,11 @@ contract AddressesMantle {
 
     /** DEX CONFIG */
 
-    uint24 internal DEX_FEE_STABLES = 100;
-    uint24 internal DEX_FEE_LOW_MEDIUM = 2500;
-    uint24 internal DEX_FEE_LOW_HIGH = 3000;
-    uint24 internal DEX_FEE_LOW = 500;
-    uint24 internal DEX_FEE_NONE = 0;
+    uint16 internal DEX_FEE_STABLES = 100;
+    uint16 internal DEX_FEE_LOW_MEDIUM = 2500;
+    uint16 internal DEX_FEE_LOW_HIGH = 3000;
+    uint16 internal DEX_FEE_LOW = 500;
+    uint16 internal DEX_FEE_NONE = 0;
 
     uint16 internal BIN_STEP_LOWEST = 1;
     uint16 internal BIN_STEP_LOW = 10;
